requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.0
entrypoints==0.3
factory-boy==3.2.1
Faker==11.1.0
flake8==3.7.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==3.0.6
pytest==6.2.5
pytest-django==3.9.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
